Move ShouldUpdatePR to evaluate to match ShouldMergePR for reuse
